+++ src/main/java/com/soorya/webscrapers/XmlParse.java
@@ -3,7 +3,6 @@ package com.soorya.webscrapers;
 import java.io.IOException;
 import java.net.URL;
 
-import org.jsoup.*;
 import org.w3c.dom.NodeList;
 import org.w3c.dom.Element;
 import org.w3c.dom.Document;
@@ -25,39 +24,25 @@ public class XmlParse
 
         document.getDocumentElement().normalize();
 
+        String htmlUrl = "";
         NodeList nodeList=document.getElementsByTagName("*");
         for (int i=0; i<nodeList.getLength(); i++)
         {
             Element element = (Element)nodeList.item(i);
             if(element.getTagName().equals("link"))
-                System.out.println(element.getTextContent());
+            {
+                htmlUrl = element.getTextContent();
+                URL urlHtml = new URL(htmlUrl);
+                new SiteScraper().siteScrape(urlHtml);
+            }
         }
+
+
     }
     public static void main(String args[]) throws IOException, ParserConfigurationException, SAXException
     {
-        //URL url = new URL("");
-        //new XmlParse().getURLfromXML(url);
-
-        URL url = new URL("http://www.thehindubusinessline.com/info-tech/fintech-involves-anothers-money-it-has-to-be-secure-paypal-ceo/article9966742.ece?homepage=true");
-
-        //returns the url in a string
-        //System.out.println(url.toExternalForm());
-
-        new SiteScraper().siteScrape(url);
-        //org.jsoup.nodes.Document doc = Jsoup.connect("https://kotaku.com/star-wars-battlefront-ii-the-kotaku-review-1820477183").get();
-        //String title = doc.title();
-        //System.out.println(title);
-
-        /*String author = doc.select(".meta__byline").first().text();
-        String time = doc.select(".meta__time").first().text();
-        String tags = doc.select(".post-tags-container").first().text();
-        String content = doc.select(".post-content").first().text();
-
-        System.out.println(author);
-        System.out.println(time);
-        System.out.println(tags);
-        System.out.println(content);
-        //System.out.println(doc.body().text());
-        */
+        URL url = new URL("https://www.anandtech.com/rss/");
+        new XmlParse().getURLfromXML(url);
+        //new SiteScraper().siteScrape(url);
     }
 }