+++ CHANGELOG.md
@@ -1,5 +1,14 @@
-# 18.11.17
+# 20.11.17
 
 ### commit
+* added class News
 
+### commit cd10d67f3722eeca34de38f5f1c5606a508afd19
+* added class SiteScraper
+* added siteScrape()
+* siteScrape() fetches content from kinja sites and a few other domains
+
+# 18.11.17
+
+### commit 0a77b258d6e33fc603b9ac3e03bc4169df7f2ce8
 * added getURLfromXML()